Move envelope2wkb() to EWKBUtils
