+++ h2/src/main/org/h2/util/geometry/EWKBUtils.java
@@ -10,6 +10,10 @@ import static org.h2.util.geometry.GeometryUtils.DIMENSION_SYSTEM_XYZ;
 import static org.h2.util.geometry.GeometryUtils.DIMENSION_SYSTEM_XYZM;
 import static org.h2.util.geometry.GeometryUtils.GEOMETRY_COLLECTION;
 import static org.h2.util.geometry.GeometryUtils.LINE_STRING;
+import static org.h2.util.geometry.GeometryUtils.MAX_X;
+import static org.h2.util.geometry.GeometryUtils.MAX_Y;
+import static org.h2.util.geometry.GeometryUtils.MIN_X;
+import static org.h2.util.geometry.GeometryUtils.MIN_Y;
 import static org.h2.util.geometry.GeometryUtils.MULTI_LINE_STRING;
 import static org.h2.util.geometry.GeometryUtils.MULTI_POINT;
 import static org.h2.util.geometry.GeometryUtils.MULTI_POLYGON;
@@ -448,6 +452,51 @@ public final class EWKBUtils {
         target.addCoordinate(source.readCoordinate(), source.readCoordinate(),
                 useZ ? source.readCoordinate() : Double.NaN, useM ? source.readCoordinate() : Double.NaN, //
                 index, total);
+    }
+
+    /**
+     * Converts an envelope to a WKB.
+     *
+     * @param envelope
+     *            envelope, or null
+     * @return WKB, or null
+     */
+    public static byte[] envelope2wkb(double[] envelope) {
+        if (envelope == null) {
+            return null;
+        }
+        byte[] result;
+        double minX = envelope[MIN_X], maxX = envelope[MAX_X], minY = envelope[MIN_Y], maxY = envelope[MAX_Y];
+        if (minX == maxX && minY == maxY) {
+            result = new byte[21];
+            result[4] = POINT;
+            Bits.writeLong(result, 5, Double.doubleToRawLongBits(minX));
+            Bits.writeLong(result, 13, Double.doubleToRawLongBits(minY));
+        } else if (minX == maxX || minY == maxY) {
+            result = new byte[41];
+            result[4] = LINE_STRING;
+            result[8] = 2;
+            Bits.writeLong(result, 9, Double.doubleToRawLongBits(minX));
+            Bits.writeLong(result, 17, Double.doubleToRawLongBits(minY));
+            Bits.writeLong(result, 25, Double.doubleToRawLongBits(maxX));
+            Bits.writeLong(result, 33, Double.doubleToRawLongBits(maxY));
+        } else {
+            result = new byte[93];
+            result[4] = POLYGON;
+            result[8] = 1;
+            result[12] = 5;
+            Bits.writeLong(result, 13, Double.doubleToRawLongBits(minX));
+            Bits.writeLong(result, 21, Double.doubleToRawLongBits(minY));
+            Bits.writeLong(result, 29, Double.doubleToRawLongBits(minX));
+            Bits.writeLong(result, 37, Double.doubleToRawLongBits(maxY));
+            Bits.writeLong(result, 45, Double.doubleToRawLongBits(maxX));
+            Bits.writeLong(result, 53, Double.doubleToRawLongBits(maxY));
+            Bits.writeLong(result, 61, Double.doubleToRawLongBits(maxX));
+            Bits.writeLong(result, 69, Double.doubleToRawLongBits(minY));
+            Bits.writeLong(result, 77, Double.doubleToRawLongBits(minX));
+            Bits.writeLong(result, 85, Double.doubleToRawLongBits(minY));
+        }
+        return result;
     }
 
     private EWKBUtils() {
